requirements.txt
django[asyncio, argon2]
django-seed
djangorestframework
djangorestframework-simplejwt
requests
bleach